chore: separate feature auth
