+++ src/features/Auth/components/Login.tsx
@@ -15,7 +15,7 @@ import {
 } from "@mui/material";
 import { Link, useNavigate } from "react-router-dom";
 import { Visibility, VisibilityOff } from "@mui/icons-material";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../../../context/AuthContext";
 
 const Login = () => {
   const [formData, setFormData] = useState({
