+++ src/features/Auth/components/SignUp.tsx
@@ -13,7 +13,7 @@ import {
 } from "@mui/material";
 import { Link, useNavigate } from "react-router-dom";
 import { Visibility, VisibilityOff } from "@mui/icons-material";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../../../context/AuthContext";
 
 const SignUp = () => {
   const [formData, setFormData] = useState({
