--- src/pages/Blog.tsx
@@ -1,5 +1,0 @@
-const Blog = () => {
-  return <div>Blog</div>;
-};
-
-export default Blog;
